Profile photo added when there are no spaces
