+++ lib/modules/spaces/pages/spaces/screens/empty_spaces.screen.dart
@@ -21,10 +21,10 @@ import 'dart:io';
 
 import 'package:demos_app/config/custom-icons/demos_icons_icons.dart';
 import 'package:demos_app/config/themes/main_theme.dart';
+import 'package:demos_app/modules/spaces/pages/spaces/widgets/go_to_profile.widget.dart';
 import 'package:demos_app/widgets/buttons/big_button_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:demos_app/config/routes/routes.dart';
-import 'package:demos_app/modules/spaces/pages/spaces/widgets/popup_spaces_menu_options.widget.dart';
 import 'package:demos_app/widgets/wrappers/safe_widget/safe_widget_validator.dart';
 
 class EmptySpacesScreen extends StatelessWidget {
@@ -34,47 +34,53 @@ class EmptySpacesScreen extends StatelessWidget {
   Widget build(BuildContext context) {
     Size size = MediaQuery.of(context).size;
     return Scaffold(
-        backgroundColor: primaryColor,
-        appBar: AppBar(
-          centerTitle: false,
-          title: const Text('Demos'),
-          actions: [PopupSpacesMenuOptions()],
-        ),
-        body: Center(
-          child: Padding(
-            padding:
-                EdgeInsets.only(
-                  bottom: Platform.isIOS ? 40.0 : 20.0, right: 24.0, left: 24.0),
-            child: Column(
-              children: [
-                Expanded(
-                  child: Column(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: [
-                      Container(
-                        margin: const EdgeInsets.only(bottom: 32.0),
-                        child: Icon(
-                          DemosIcons.empty_spaces,
-                          color: primaryColorLight,
-                          size: size.width * 0.2,
-                        ),
+      backgroundColor: primaryColor,
+      appBar: AppBar(
+        centerTitle: false,
+        title: const Text('Demos'),
+        actions: const [GoToProfile()],
+      ),
+      body: Center(
+        child: Padding(
+          padding: EdgeInsets.only(
+            bottom: Platform.isIOS ? 40.0 : 20.0,
+            right: 24.0,
+            left: 24.0,
+          ),
+          child: Column(
+            children: [
+              Expanded(
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: [
+                    Container(
+                      margin: const EdgeInsets.only(bottom: 32.0),
+                      child: Icon(
+                        DemosIcons.empty_spaces,
+                        color: primaryColorLight,
+                        size: size.width * 0.2,
                       ),
-                      const Text('No tienes espacios por el momento.',
-                          textAlign: TextAlign.center,
-                          style: TextStyle(
-                              color: primaryColorLight, fontSize: 16.0)),
-                    ],
-                  ),
+                    ),
+                    const Text(
+                      'No tienes espacios por el momento.',
+                      textAlign: TextAlign.center,
+                      style:
+                          TextStyle(color: primaryColorLight, fontSize: 16.0),
+                    ),
+                  ],
                 ),
-                SafeWidgetValidator(
-                    child: BigButton(
+              ),
+              SafeWidgetValidator(
+                child: BigButton(
                   onPressed: () => goToCreateNewSpace(context),
                   text: 'Crea tu primer espacio',
-                ))
-              ],
-            ),
+                ),
+              )
+            ],
           ),
-        ));
+        ),
+      ),
+    );
   }
 
   void goToCreateNewSpace(BuildContext context) {
